requirements.txt
decorator==3.4.2
Flask==0.10.1
Flask-SQLAlchemy==2.0
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
pbr==0.11.0
requests==2.6.2
six==1.9.0
SQLAlchemy==0.9.9
sqlalchemy-migrate==0.9.6
sqlparse==0.1.15
Tempita==0.5.2
Werkzeug==0.10.4
